extracts timestamp logic into rules/ folder and allows easy expansion of further business rules
Also tests logging behaviour
